Combine Dev Shells & Fix Pure Build (#1761)
* combined dev shells and fix for unlocated libffi.h

SDL2 is not very heavy to include for all users.  This way, no switching shells
is necessary to build Lem or LemSDL2

pkg-config is a requirement.

Including packages in the LD_LIBRARY_PATH list does not appear to run the
pkg-config hook, so anything where pkg-config is run needs to be in the package list.

* which is needed inside completely clean builds

Most systems already have this, but it's pretty cheap to make sure.  Not relying
on any incidental resources is the Nix way.